requirements.txt
fastapi~=0.112.1
uvicorn~=0.30.6
sqlalchemy~=2.0.34
pydantic~=2.8.2
passlib~=1.7.4
pyjwt~=2.9.0
alembic~=1.13.2
pydantic-settings~=2.4.0
websockets~=12.0
redis~=5.1.1